Conflicts:
	GameDemo-android/assets/spine/ted/ted.json
	GameDemo-android/assets/spine/ted/ted.skel

+++ GameDemo/src/com/angrynerds/game/screens/play/PlayController.java
@@ -46,7 +46,7 @@ public class PlayController {
      */
     public void update(float deltaTime) {
         world.update(deltaTime);
-        controllerUI.getLifeBar().setLifePercent((float)world.getPlayer().getActualHP()/(float)world.getPlayer().getMaxHP());
+        controllerUI.getLifeBar().setLifePercent(world.getPlayer().getActualHP()/world.getPlayer().getMaxHP());
     }
 
     /**
